# openapi.yaml
openapi: 3.0.3
info:
  title: DogsOut API
  description: Backend API for the DogsOut mobile app
  version: 0.0.1

servers:
  - url: http://localhost:8080
    description: Local development

security:
  - bearerAuth: []

paths:
  /auth/register:
    post:
      summary: Register a new user
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              email: max@example.com
              name: Max Mustermann
              password: securepass123
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error (invalid email, password too short, etc.)
        '409':
          description: Email already in use

  /auth/login:
    post:
      summary: Login with email and password
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: max@example.com
              password: securepass123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
        '401':
          description: Invalid credentials

  /auth/google:
    post:
      summary: Sign in with Google
      description: Verify a Google ID token obtained from the Google Sign-In SDK and return a JWT.
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleAuthRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
        '401':
          description: Invalid or expired Google token

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    RegisterRequest:
      type: object
      required: [email, name, password]
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        password:
          type: string
          minLength: 8

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    GoogleAuthRequest:
      type: object
      required: [idToken]
      properties:
        idToken:
          type: string
          description: Google ID token obtained from the Google Sign-In SDK

    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT to include in the Authorization header of subsequent requests
        email:
          type: string
        name:
          type: string